+++ AdminSite/src/adminsite/adminsite/site/employees/timesheet/timesheetdata.aspx.cs
@@ -27,7 +27,7 @@ namespace adminsite.site.employees.timesheet
                         string timesheetString = (string)Session["CONSULTED_TIMESHEET"];
                         if (timesheetString != null)
                         {
-                            Timesheet timesheet = new Timesheet(Int32.Parse(timesheetString));
+                            Timesheet timesheet = new Timesheet(Int64.Parse(timesheetString));
                             GetAllWorkloadsByTimesheetCommand cmd = new GetAllWorkloadsByTimesheetCommand(timesheet);
                             cmd.Execute();
                             timesheet = cmd.GetResults();
@@ -201,7 +201,7 @@ namespace adminsite.site.employees.timesheet
             try
             {
                 string timesheetString = (string)HttpContext.Current.Session["CONSULTED_TIMESHEET"];
-                Timesheet timesheet = new Timesheet(Int32.Parse(timesheetString));
+                Timesheet timesheet = new Timesheet(Int64.Parse(timesheetString));
                 GetAllWorkloadsByTimesheetCommand cmd = new GetAllWorkloadsByTimesheetCommand(timesheet);
                 cmd.Execute();
                 timesheet = cmd.GetResults();
